module-0/studio/requirements.txt
langchain_openai 
langchain_core 
langchain_community 
tavily-python
python-dotenv